video-py3/requirements.txt
requests
oss2
opencv-python
numpy
pillow
